custom metric - trying to overcome Tensor AttribureError

+++ task_code/metrics.py
@@ -15,4 +15,4 @@ def pix_to_pix_rmse(img1, img2):
     error = img1.flatten() - img2.flatten()
     square_error = np.multiply(error, error)
     mean_square_error = square_error.mean()
-    return np.sqrt(mean_square_error)
+    return float(np.sqrt(mean_square_error))
